+++ src/consumer/component/package-json.js
@@ -1,6 +1,8 @@
 /** @flow */
 
 import R from 'ramda';
+import path from 'path';
+import fs from 'fs-extra';
 import format from 'string-format';
 import { BitId } from '../../bit-id';
 import Component from '../component';
@@ -10,7 +12,8 @@ import {
   DEFAULT_REGISTRY_DOMAIN_PREFIX,
   DEFAULT_SEPARATOR,
   ASTERISK,
-  COMPONENTES_DEPENDECIES_REGEX
+  COMPONENTES_DEPENDECIES_REGEX,
+  NODE_PATH_SEPARATOR
 } from '../../constants';
 import ComponentMap from '../bit-map/component-map';
 import { filterAsync, pathNormalizeToLinux, pathRelative } from '../../utils';
@@ -38,6 +41,21 @@ async function addComponentsToRoot(consumer: Consumer, components: Component[]) 
   );
   const registryPrefix = getRegistryPrefix();
   await PackageJson.addComponentsIntoExistingPackageJson(consumer.getPath(), componentsToAdd, registryPrefix);
+}
+
+/**
+ * Add given components with their versions to root package.json
+ */
+async function addComponentsWithVersionToRoot(consumer: Consumer, componentsIds: BitId[]) {
+  const driver = await consumer.driver.getDriver(false);
+  const PackageJson = driver.PackageJson;
+
+  const componentsToAdd = R.fromPairs(
+    componentsIds.map((id) => {
+      return [id.toStringWithoutVersion(), id.version];
+    })
+  );
+  await PackageJson.addComponentsIntoExistingPackageJson(consumer.getPath(), componentsToAdd, getRegistryPrefix());
 }
 
 /**
@@ -95,8 +113,14 @@ async function changeDependenciesToRelativeSyntax(
   return Promise.all(components.map(component => updateComponent(component)));
 }
 
-function getRegistryPrefix() {
+function getRegistryPrefix(): string {
   return getSync(CFG_REGISTRY_DOMAIN_PREFIX) || DEFAULT_REGISTRY_DOMAIN_PREFIX;
+}
+
+function convertIdToNpmName(id: BitId, withVersion = false): string {
+  const registryPrefix = getRegistryPrefix();
+  const npmName = `${registryPrefix}/${id.toStringWithoutVersion().replace(/\//g, NODE_PATH_SEPARATOR)}`;
+  return withVersion ? `${npmName}@${id.version}` : npmName;
 }
 
 async function write(
@@ -120,7 +144,7 @@ async function write(
   const registryPrefix = getRegistryPrefix();
   const name = excludeRegistryPrefix
     ? component.id.toStringWithoutVersion().replace(/\//g, '.')
-    : `${registryPrefix}/${component.id.toStringWithoutVersion().replace(/\//g, '.')}`;
+    : convertIdToNpmName(component.id);
   const packageJson = new PackageJson(bitDir, {
     name,
     version: component.version,
@@ -188,11 +212,23 @@ async function addWorkspacesToPackageJson(
   }
 }
 
+async function removeComponentsFromNodeModules(consumer: Consumer, componentIds: BitId[]) {
+  const registryPrefix = getRegistryPrefix();
+  // paths without scope name, don't have a symlink in node-modules
+  const pathsToRemove = componentIds
+    .map((id) => {
+      return id.scope ? Consumer.getNodeModulesPathOfComponent(registryPrefix, id) : null;
+    })
+    .filter(a => a); // remove null
+
+  return Promise.all(pathsToRemove.map(componentPath => fs.remove(path.join(consumer.getPath(), componentPath))));
+}
+
 async function removeComponentsFromWorkspacesAndDependencies(
   consumer: Consumer,
   rootDir: string,
   bitMap: BitMap,
-  componentIds: string[]
+  componentIds: BitId[]
 ) {
   const driver = await consumer.driver.getDriver(false);
   const PackageJson = driver.PackageJson;
@@ -209,12 +245,14 @@ async function removeComponentsFromWorkspacesAndDependencies(
     getRegistryPrefix(),
     componentIds.map(id => id.toStringWithoutVersion())
   );
+  await removeComponentsFromNodeModules(consumer, componentIds);
 }
 
 export {
   addComponentsToRoot,
   changeDependenciesToRelativeSyntax,
   write,
+  addComponentsWithVersionToRoot,
   updateAttribute,
   addWorkspacesToPackageJson,
   removeComponentsFromWorkspacesAndDependencies