+++ src/scope/exceptions/cyclic-dependencie.js
@@ -1,0 +1,8 @@
+/** @flow */
+export default class CyclicDependencies extends Error {
+  msg: string;
+  constructor(msg: string) {
+    super();
+    this.msg = msg;
+  }
+}